build(docker): fix XRootD repository location (#95)
The CERN storage team repository from which the XRootD packages were
installed has temporarily moved its URL location from "debian" to
"old_debian", with plans to move it back. This fix therefore attempts to
recognise both URLs so that the builds would work regardless of where
the repository is hosted.